+++ packages/common/src/index.ts
@@ -19,4 +19,6 @@ export * from "./utils/sse.js";
 export * from "./utils/timeZone.js";
 export * from "./schemas/workspaceSettings.schema.js";
 export * from "./schemas/userSettings.schema.js";
+export * from "./schemas/asset.schema.js";
+export * from "./schemas/ad.schema.js";
 export * from "./i18n/apiMessages.js";